requirements.txt
asgiref==3.4.1
astroid==2.8.5
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.3.1
django-modeltranslation==0.16.1
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.3
pylint==2.11.1
python-decouple==3.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.0
whitenoise==6.0.0
wrapt==1.13.3
